+++ backend/src/main/java/com/tushar/split/Model/Users.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class Users {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     @Column(unique = true,nullable = false)
     private String username;
     @Column(unique = true,nullable = false)
@@ -27,8 +27,6 @@ public class Users {
     @ManyToMany(mappedBy = "members")
     @JsonIgnore
     private List<SplitGroups> groups;
-
-
     @OneToMany(mappedBy = "paidBy")
     @JsonIgnore
     private List<Expense> expenses;
