+++ backend/src/main/java/com/tushar/split/Repo/OtpRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface OtpRepo extends JpaRepository<OtpVerification,Long> {
+public interface OtpRepo extends JpaRepository<OtpVerification,Integer> {
 
     Optional<OtpVerification> findByEmail(String email);
 }
